+++ src/unit.tcl
@@ -295,9 +295,7 @@ proc unit::TemplateData {wtxt l1 l2 tpldata} {
   set sec [clock seconds]
   set fname [alited::bar::FileName]
   # fill the common wildcards
-  set abra {*^e!`i@U50=|}
-  set tex [string map [list %% $abra] $tex]
-  set tex [string map [list \
+  set tex [alited::Map {} $tex \
     %d [alited::tool::FormatDate $sec] \
     %t [clock format $sec -format $al(TPL,%t) -locale $::alited::al(LOCAL)] \
     %u $al(TPL,%u) \
@@ -309,8 +307,7 @@ proc unit::TemplateData {wtxt l1 l2 tpldata} {
     %n [file rootname [file tail $fname]] \
     %A $DIR \
     %M $al(EM,mnudir) \
-    ] $tex]
-  set tex [string map [list $abra %] $tex]
+    ]
   # get a list of proc/method's arguments:
   # from "proc pr {ar1 ar2 ar3} " and a template "  # %a -\n"
   # to get